Moved controllers into a separate folder [Jerry Chang]
