+++ src/angular/controllers/navbar.controller.js
@@ -1,0 +1,54 @@
+var controllers = angular.module('controllers');
+
+controllers.controller('NavbarCtrl', ['$scope', function ($scope) {}]);
+
+controllers.controller('ServiceCtrl', ['$scope', '$routeParams', 'Search', function($scope, $routeParams, Search) {
+  var servicesById = Search.getServicesById();
+  if(servicesById) {
+        var service = servicesById[$routeParams.serviceId];
+        $scope.service = {};
+        $scope.service.id = service.id;
+        $scope.service.locationName = service.properties.locationName;
+        $scope.service.partnerName = service.properties.partnerName;
+        $scope.service.comments = service.properties.comments;
+        $scope.service.activityCategory = service.properties.activityCategory;
+        $scope.service.activityName = service.properties.activityName;
+        $scope.service.startDate = service.properties.startDate;
+        $scope.service.endDate = service.properties.endDate;
+        $.each(service.properties.indicators, function(index, value) {
+            if(value) {
+                $scope.service.activityDetails = index;
+            }
+        });
+        var propList = new Array();
+        $scope.hours = new Array();
+        $.each(service.properties, function(index, value) {
+            var tempArray = new Array();
+            tempArray = index.split(".");
+            if(index != 'comments' && tempArray.length > 1) {
+                if($.isNumeric(tempArray[0])) {
+                    //TODO: Let's see if we can print it from index rather than creating new object for it again.
+                    var obj = {};
+                    var level = parseInt(tempArray[0], 10);
+                    if(level != 8) {
+                        obj.key = $.trim(tempArray[1]);
+                        $.each(service.properties[index], function(index, value) {
+                            if(value) {
+                                obj.value = index;
+                            }
+                        });
+                        propList[level] = obj;
+                    } else {
+                        $.each(service.properties[index], function(index, value) {
+                            if(value) {
+                                $scope.hours.push(index);
+                            }
+                        });
+                    }
+                }
+            }
+        });
+        propList = $.grep(propList, function(n){ return(n) });
+        $scope.service.properties = propList;
+    }
+}]);
