+++ src/angular/controllers/search.controller.js
@@ -1,0 +1,76 @@
+var controllers = angular.module('controllers');
+
+/**
+ * For the category/region search view
+ */
+controllers.controller('SearchCtrl', ['$scope', '$http', '$location', 'ServicesList', 'Search', function ($scope, $http, $location, ServicesList, Search) {
+
+    ServicesList.get(function (data) {
+        $scope.services = data;
+        // Here we're going to extract the list of categories and display them in a simple template
+        
+        // use an object to collect service information since object keys won't allow
+        // for duplicates (this basically acts as a set)
+        var categories = {};
+        var regions = {};
+        var servicesById = {};
+        $.each(data, function (index, service) {
+            // add activity and its category to list, and increment counter of this category's available services
+            var category = service.properties.activityCategory;
+            if (category) {
+                if (categories[category] == null) {
+                    categories[category] = {activities:{}, count: 0};
+                }
+                categories[category].count++;
+
+                var activity = service.properties.activityName;
+                if (activity) {
+                    if (categories[category].activities[activity] == null) {
+                        categories[category].activities[activity] = {name: activity, count: 0};
+                    }
+                    categories[category].activities[activity].count++;
+                }
+            }
+            
+            // add region to list, and increment counter of this region's available services
+            var region = service.properties.locationName;
+            if (region) {
+                if (regions[region] == null) {
+                    regions[region] = 0;
+                }
+                regions[region]++;
+            }
+
+            var id = service.id;
+      if(id) {
+        if(servicesById[id] == null) {
+          servicesById[id] = service;
+        }
+//        servicesById[id]++;
+      }
+        });
+
+        // now to get an array of categories we just map over the keys of the object
+        $scope.categories = $.map(categories, function (value, index) {
+            return {name: index, count: value.count, activities: value.activities};
+        });
+
+        // now to get an array of regions we just map over the keys of the object
+        $scope.regions = $.map(regions, function (value, index) {
+            return {name: index, count: value};
+        });
+
+        $scope.services = $.map(servicesById, function (value, index) {
+            return {id: index, service: value};
+        });
+        Search.setServicesById(servicesById);
+    });
+
+    /**
+     * When the user clicks on a category, we filter by that category and then navigate to the results view
+     */
+    $scope.selectCategory = function (category) {
+        Search.selectCategory(category);
+        $location.path("/results")
+    }
+}]);
